backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
email-validator
python-jose[cryptography]
bcrypt
openai
requests
gspread
oauth2client
python-wordpress-xmlrpc
beautifulsoup4
python-slugify
Pillow
websockets
sqlalchemy[asyncio]
asyncpg
alembic
cryptography
